requirements.txt
langchain
streamlit
google_generativeai
python_dotenv